refactor(storage): move archive helper into storage module
